backend/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
pydantic
python-multipart
polyline
firebase-admin
python-dotenv
psycopg2-binary
Pillow
google-cloud-storage
